# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/explanation-detail": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "公共方法"
                ],
                "summary": "题解详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "explanation_identity",
                        "name": "explanation_identity",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":\"200\",\"data\":\"\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/explanation-list": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "公共方法"
                ],
                "summary": "题解列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "problem_identity",
                        "name": "problem_identity",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "size",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":\"200\",\"data\":\"\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/explanation-readNum": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "公共方法"
                ],
                "summary": "题解阅读量",
                "parameters": [
                    {
                        "type": "string",
                        "description": "explanation_identity",
                        "name": "explanation_identity",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":\"200\",\"data\":\"\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/reply": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "公共方法"
                ],
                "summary": "回复列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "comment_identity",
                        "name": "comment_identity",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "size",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":\"200\",\"data\":\"\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/user/explanation-create": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "用户私有方法"
                ],
                "summary": "题解创建",
                "parameters": [
                    {
                        "type": "string",
                        "description": "authorization",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "ExplanationCreate",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/define.ExplanationCreate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":\"200\",\"data\":\"\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/explanation-delete": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "用户私有方法"
                ],
                "summary": "题解删除",
                "parameters": [
                    {
                        "type": "string",
                        "description": "authorization",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "user_identity",
                        "name": "user_identity",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "explanation_identity",
                        "name": "explanation_identity",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":\"200\",\"data\":\"\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/explanation-likeNum": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "用户私有方法"
                ],
                "summary": "题解点赞",
                "parameters": [
                    {
                        "type": "string",
                        "description": "authorization",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "explanation_identity",
                        "name": "explanation_identity",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":\"200\",\"data\":\"\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/explanation-modify": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "用户私有方法"
                ],
                "summary": "题解修改",
                "parameters": [
                    {
                        "type": "string",
                        "description": "authorization",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "ExplanationModify",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/define.ExplanationModify"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":\"200\",\"data\":\"\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/reply-create": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "用户私有方法"
                ],
                "summary": "回复创建",
                "parameters": [
                    {
                        "type": "string",
                        "description": "authorization",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "ReplyCreate",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/define.ReplyCreate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":\"200\",\"data\":\"\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/reply-delete": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "用户私有方法"
                ],
                "summary": "回复删除",
                "parameters": [
                    {
                        "type": "string",
                        "description": "authorization",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "user_identity",
                        "name": "user_identity",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "reply_identity",
                        "name": "reply_identity",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":\"200\",\"data\":\"\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "define.ExplanationCreate": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "problemIdentity": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "define.ExplanationModify": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "identity": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "userIdentity": {
                    "type": "string"
                }
            }
        },
...
        "define.ReplyCreate": {
            "type": "object",
            "properties": {
                "CommentIdentity": {
                    "type": "string"
                },
                "Reply": {
                    "type": "string"
                }
            }
        },
...
    }
}